Error logging and structure changes
* Moved case class/objects to separate files. This improves clarity and structure.
* Fixed the use of `null` in comparison expressions. The solution was to wrap the `null` in
an `Option` and then checking if the object `isDefined`.
* Implemented logging with typesafe/scalalogging library
* Upped scala-async version to 0.9.5
* Heavy code style changes. Please see [https://github.com/databricks/scala-style-guide]
