+++ build.sbt
@@ -21,12 +21,14 @@ parallelExecution in Test := false
 val playVersion = "2.4.3"
 
 libraryDependencies ++= Seq(
+  "ch.qos.logback" % "logback-classic"  % "1.1.3",
   "org.scalatest" %% "scalatest" % "2.2.4" % "test",
   "com.typesafe.play" %% "play-json" % playVersion,
   "com.typesafe.play" %% "play-ws" % playVersion,
   "com.typesafe.play" %% "play-iteratees" % playVersion,
   "com.typesafe.play.extras" %% "iteratees-extras" % "1.5.0",
-  "org.scala-lang.modules" %% "scala-async" % "0.9.2"
+  "com.typesafe.scala-logging" %% "scala-logging"  % "3.1.0",
+  "org.scala-lang.modules" %% "scala-async" % "0.9.5"
 )
 
 seq(lsSettings :_*)
