+++ src/main/scala/org/anormcypher/Utils.scala
@@ -2,7 +2,9 @@ package org.anormcypher
 
 import scala.language.implicitConversions
 
-case class MayErr[+E, +A](e: Either[E, A]) {
+import com.typesafe.scalalogging.StrictLogging
+
+case class MayErr[+E, +A](e: Either[E, A]) extends StrictLogging {
 
   def flatMap[B, EE >: E](f: A => MayErr[EE, B]): MayErr[EE, B] = {
     MayErr(e.right.flatMap(a => f(a).e))
@@ -12,16 +14,19 @@ case class MayErr[+E, +A](e: Either[E, A]) {
     MayErr(e.right.map(f))
   }
 
-  def filter[EE >: E](p: A => Boolean, error: EE): MayErr[EE, A] = MayErr(e.right.filter(p).getOrElse(Left(error)))
+  def filter[EE >: E](p: A => Boolean, error: EE): MayErr[EE, A] =
+    MayErr(e.right.filter(p).getOrElse(Left(error)))
 
   def toOptionLoggingError(): Option[A] = {
-    e.left.map(m => { println(m.toString); m }).right.toOption
+    e.left.map(m => { logger.info(m.toString); m }).right.toOption
   }
 
-  def get = e.fold(e => throw new RuntimeException(e.toString), a => a)
+  def get: A = e.fold(e => throw new RuntimeException(e.toString),a => a)
+
 }
 
 object MayErr {
   implicit def eitherToError[E, EE >: E, A, AA >: A](e: Either[E, A]): MayErr[EE, AA] = MayErr[E, A](e)
+
   implicit def errorToEither[E, EE >: E, A, AA >: A](e: MayErr[E, A]): Either[EE, AA] = e.e
 }
