+++ src/test/scala/org/anormcypher/BaseAnormCypherSpec.scala
@@ -1,24 +1,45 @@
 package org.anormcypher
 
-import org.scalatest._
-import play.api.libs.ws._, ning._
-import scala.concurrent._
+import scala.concurrent.ExecutionContext
 
-trait BaseAnormCypherSpec extends FlatSpec with Matchers with BeforeAndAfterEach with BeforeAndAfterAll {
+import org.scalatest.BeforeAndAfterAll
+import org.scalatest.BeforeAndAfterEach
+import org.scalatest.FlatSpec
+import org.scalatest.Matchers
+import org.scalatest.concurrent.ScalaFutures
+import org.scalatest.time.Minutes
+import org.scalatest.time.Span
+
+import play.api.libs.ws.ning.NingWSClient
+
+import com.typesafe.scalalogging.StrictLogging
+
+trait BaseAnormCypherSpec
+  extends FlatSpec
+  with Matchers
+  with BeforeAndAfterEach
+  with BeforeAndAfterAll
+  with StrictLogging {
+  
   val wsclient = NingWSClient()
-  implicit val neo4jrest = Neo4jREST(scala.util.Properties.envOrElse("NEO4J_SERVER", "localhost"))(wsclient)
+
+  implicit val neo4jrest =
+    Neo4jREST(scala.util.Properties.envOrElse("NEO4J_SERVER", "localhost"))(wsclient)
+
   implicit val ec = ExecutionContext.global
 
-  override def afterAll = {
-    wsclient.close()
-  }
+  override def afterAll: Unit = wsclient.close()
+
 }
 
 package async {
-  import org.scalatest.concurrent.ScalaFutures
 
-  trait BaseAsyncSpec extends BaseAnormCypherSpec with ScalaFutures {
-    import org.scalatest.time._
-    implicit override val patienceConfig = PatienceConfig(timeout = Span(2, Minutes))
+  trait BaseAsyncSpec
+    extends BaseAnormCypherSpec
+    with ScalaFutures {
+
+    implicit override val patienceConfig =
+      PatienceConfig(timeout = Span(2, Minutes))
   }
+
 }
