+++ src/test/scala/org/anormcypher/CypherParserTest.scala
@@ -3,9 +3,16 @@ package org.anormcypher
 import CypherParser._
 
 class CypherParserSpec extends BaseAnormCypherSpec {
+
+  val getTestNodeCountQuery = """
+    START n = node(*)
+    WHERE n.tag = 'anormcyphertest'
+    RETURN count(n)
+  """
+
   override def beforeEach = {
     // initialize some test data
-    Cypher("""create 
+    Cypher("""CREATE 
       (us {type:"Country", name:"United States", code:"USA", tag:"anormcyphertest"}),
       (germany {type:"Country", name:"Germany", code:"DEU", population:81726000, tag:"anormcyphertest"}),
       (france {type:"Country", name:"France", code:"FRA", tag:"anormcyphertest", indepYear:1789}),
@@ -28,10 +35,10 @@ class CypherParserSpec extends BaseAnormCypherSpec {
 
   override def afterEach = {
     // delete the test data
-    Cypher("""match (n)     
-      where n.tag = "anormcyphertest"
-      optional match (n)-[r]-()
-      delete n, r;
+    Cypher("""MATCH (n)     
+      WHERE n.tag = "anormcyphertest"
+      OPTIONAL MATCH (n)-[r]-()
+      DELETE n, r;
       """)()
   }
 
@@ -45,26 +52,19 @@ class CypherParserSpec extends BaseAnormCypherSpec {
   }
 
   it should "be able to parse into a single Long" in {
-    val count: Long = Cypher("""
-      start n=node(*) 
-      where n.tag = 'anormcyphertest' 
-      return count(n)""").as(scalar[Long].single)
+    val count: Long = Cypher(getTestNodeCountQuery).as(scalar[Long].single)
     count should equal (11)
   }
 
   it should "be able to parse into a single Float" in {
-    val count: Float = Cypher("""
-      start n=node(*)
-      where n.tag = 'anormcyphertest'
-      return count(n)""").as(scalar[Float].single)
+    val count: Float = Cypher(getTestNodeCountQuery).as(scalar[Float].single)
     count should equal (11)
   }
 
   it should "be able to parse a case class with a node" in {
     val results = Cypher("start n=node(*) where n.type = 'Country' return n.name as name, n")().map {
       case CypherRow(name: String, n: NeoNode) => name -> n
-      case e:Any => {//println(e);
-      }
+      case e:Any => logger.info(s"$e")
     }.toList
     // TODO this isn't working!
     //results.head("United States").props should equal (Map("name" -> "United States", "type" -> "Country", "tag" -> "anormcyphertest", "code" -> "USA"))
@@ -73,10 +73,10 @@ class CypherParserSpec extends BaseAnormCypherSpec {
   it should "be able to parse and flatten into a tuple" in {
     val result:List[(String,Int)] = 
       Cypher("""
-        start n=node(*) 
-        where n.type = 'Country' and has(n.name) and has(n.population) 
-        return n.name, n.population 
-        order by n.name
+        START n=node(*) 
+        WHERE n.type = 'Country' AND HAS(n.name) AND HAS(n.population) 
+        RETURN n.name, n.population 
+        ORDER BY n.name
         """
       ).as(
         (str("n.name") ~ int("n.population")).map(flatten).*
